+++ examples/resources/incident_catalog_entries/resource.tf
@@ -98,10 +98,9 @@ resource "incident_catalog_entries" "services" {
       # e.g. artist-web
       name = entry["metadata"]["name"],
 
-      # In this catalog we know names are unique, so we can use them as a
-      # human-friendly unique alias. Other catalogs name may not be unique, in
-      # which case this would fail.
-      alias = entry["metadata"]["name"],
+      # We want to use the name as a potential reference value for this entry
+      # from other entries.
+      aliases = [entry["metadata"]["name"]],
 
       # Now build all attribute values for this entry, with an object
       # comprehension that filters out any attributes that we are missing values